requirements.txt
geopy==2.4.1pandas==2.2.3pyarrow==19.0.0psycopg2-binary==2.9.10sqlalchemy==2.0.37PyYAML==6.0.2